Add general->competition module and method
I am porting the example script to use th library, this
is helping me develop the code as I have a direction to follow.

First call is to the general.get_one method of the API to get the
basic competition info. I drop the get_one as it felt redundant.